Feature/ Add option for audio-follows-video (#78)
We should implement an option that can be selected in the gear menu of
the hosted player. This audio option has the following requirements:

- NOT the default option
- available only if multiple audio tracks exist
- when this option is chosen, when a user switches to a video that
contains audio the audio will change to match the video. If the video
does not contain audio track and is video-only, the last audio track
will play until a video that has matched audio is chosen.
- the user can switch back to an explicit audio track at any time and
audio will no longer follow video

How to enable this option 

- Audio Source > Audio Follows Video (top option in the list)
- Add a queryParam for audioFollowsVideo=true/false