Feature / Add startingQuality query param for layer selection (#86)
Added in the paramOptions a new option called `startingQuality` which
will set the layer specified, if any available (simulcast stream). There
are 2 formats accepted:

1. Category of quality: posible values are **High**, **Medium** or
**Low**.
2. Height of the video: a numerical value with a maximum length of 4
representing the desired height of the video. This value will be mapped
to the corresponding height layer. If the entered value is invalid or
does not match the height of any available layers, a warning will be
logged. In such cases, the Auto quality setting will be automatically
selected.

---------

Signed-off-by: Santiago Souto <ssout@dolby.com>