Fix/Firefox stats not showing all values (#84)
Change the way values are parsed when stats event is called. Before,
stats used `mid` value to select which stats to show. Now it is based on
the `trackIdentifier` of the stats event's payload, that is available in
all browsers.

On Firefox, when viewing the Media Stats, it should be shown video stats
that previously were not shown.

This fix depends on a change from the Millicast SDK.